+++ src/Concerns/ManagesTransactions.php
@@ -141,6 +141,8 @@ trait ManagesTransactions
             $this->currentTransaction->commit();
         }
 
+        $this->transactionsManager?->stageTransactions($this->getName());
+
         $this->transactions = max(0, $this->transactions - 1);
         if ($this->isTransactionFinished()) {
             $this->currentTransaction = null;