+++ CHANGELOG.md
@@ -12,6 +12,9 @@ Changed
 - [Breaking] SessionNotFoundErrorMode was removed and will always run clear session pool. (#132) (#130)
 - [Breaking] Auth cache and Session pool now share the same file cache adapter (#139)
 
+Fixed
+- Explicitly stage/clear transaction on commit to correctly run afterCommit jobs in Laravel >= [v10.32.0](https://github.com/laravel/framework/pull/48859) (#144)
+
 # v5.2.2 (2023-08-22)
 
 Fixed